ASoC: SOF: Add MFD support
Add MFD support in SOF with a single client device for audio.
The audio client is registered as a separate platform device
under the top-level SOF device. This patch deals with
partitioning the DSP hardware initialisation parts and
the audio specific code. The hardware initialization,
FW load/boot, IPC init is done at the top-level device probe.
The audio client platform device's probe deals with registering
the component driver, creating the machine driver and loading
the topology.

The device hierarchy is altered with the addition of the platform
device for the audio client. The machine driver and the component
driver will now be children of the sof-audio client platform
device.

The top-level SOF device still has some audio-related members and
ops and will be partitioned in the following patch.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>