ASoC: SOF: Add platform device for audio
Create a new platform device for audio that takes care
of registering the machine driver and the component
driver. The audio platform device is registered using the
newly defined sof_client_dev_register/unregister API's.

DSP Hardware initialization, FW load/boot and IPC init
are all handled at the top-level SOF device probe.

The device hierarchy is altered with the addition of the platform
device for the audio client. The machine driver and the component
driver will now be children of the sof-audio client platform
device.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>