ASoC: SOF: Add a prepare op to IPC topology widget ops
In order to set up a pipeline with IPC4, the total memory usage for the
pipeline needs to be calculated based on the list of connected widgets.
Add new prepare() op to struct sof_ipc_tplg_widget_ops that will be used
to calculate the moemory usage for each widget in the pipelines
associated with a PCM.

Before setting up the widgets, sof_prepare_widgets_in_path() walks
through the list of connected widgets and prepares the widgets for set
up.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>