ASoC: SOF: ipc4-compress: Use codec parameters for initial hw_params
Use the audio information from the compr_params to configure the PCM
hw_params to be used to set up the audio graph.

The codec configuration gives an estimate of the expected rate progression
of the audio stream.

Note: for the host side the format, rate, channels actually do not matter
as we are treating the compr stream as a bytestream.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>